feat(KillAura): randomized scan and swing range (#5789)
fixes https://github.com/CCBlueX/LiquidBounce/issues/5756

+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/combat/elytratarget/TargetEntityMovementPrediction.kt
@@ -18,6 +18,6 @@ internal object TargetEntityMovementPrediction : ToggleableConfigurable(ElytraRo
     ) = if (!enabled || (glidingOnly && !target.isGliding)) {
         targetPosition
     } else {
-        targetPosition + target.velocity * multiplier.random()
+        targetPosition + target.velocity * multiplier.random().toDouble()
     }
 }
